requirements.txt
APScheduler==3.5.3
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Django==2.1.3
django-apscheduler==0.2.13
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
idna==2.7
numpy==1.15.4
Pillow==5.3.0
pytz==2018.7
requests==2.20.1
six==1.11.0
tzlocal==1.5.1
urllib3==1.24.1
